finish small file upload

+++ project/src/Entities/user.go
@@ -4,6 +4,6 @@ type User struct {
 	UserId   int    `json:"id" form:"id"`
 	UserName string `json:"username" form:"username"`
 	Password []byte `json:"password" form:"password"`
-	Capacity int    `json:"capacity" form:"capacity"`
-	Margin   int    `json:"margin" form:"margin"`
+	Capacity uint64 `json:"capacity" form:"capacity"`
+	Margin   uint64 `json:"margin" form:"margin"`
 }
